use the binaryen ctor-eval tool for wasm instead of the js-based one (#5194)
* use the binaryen ctor-eval tool for wasm instead of the js-based ctor-evaller. the binaryen tool is much faster and more efficient (but does not special-case atexit, so it works best with NO_EXIT_RUNTIME; it suggests doing so if not already enabled)

* note that when interpreting the wasm, we can't use the ctor evaller, the bundled interpreter confuses the tool

* do not simplifyIfs when wasm, as it hits a bug in asm2wasm

+++ emcc.py
@@ -1305,9 +1305,6 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         #  * if we also supported js mem inits we'd have 4 modes
         #  * and js mem inits are useful for avoiding a side file, but the wasm module avoids that anyhow
         memory_init_file = True
-        if shared.Building.is_wasm_only() and shared.Settings.EVAL_CTORS:
-          logging.debug('disabling EVAL_CTORS, as in wasm-only mode it hurts more than it helps. TODO: a wasm version of it')
-          shared.Settings.EVAL_CTORS = 0
         # async compilation requires wasm-only mode, and also not interpreting (the interpreter needs sync input)
         if shared.Settings.BINARYEN_ASYNC_COMPILATION == 1 and shared.Building.is_wasm_only() and 'interpret' not in shared.Settings.BINARYEN_METHOD:
           # async compilation requires a swappable module - we swap it in when it's ready
@@ -1326,9 +1323,18 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           sys.exit(1)
 
       if shared.Settings.EVAL_CTORS:
-        # this option is not a js optimizer pass, but does run the js optimizer internally, so
-        # we need to generate proper code for that
-        shared.Settings.RUNNING_JS_OPTS = 1
+        if not shared.Settings.BINARYEN:
+          # for asm.js: this option is not a js optimizer pass, but does run the js optimizer internally, so
+          # we need to generate proper code for that (for wasm, we run a binaryen tool for this)
+          shared.Settings.RUNNING_JS_OPTS = 1
+        else:
+          if 'interpret' in shared.Settings.BINARYEN_METHOD:
+            logging.warning('disabling EVAL_CTORS as the bundled interpreter confuses the ctor tool')
+            shared.Settings.EVAL_CTORS = 0
+          else:
+            # for wasm, we really want no-exit-runtime, so that atexits don't stop us
+            if not shared.Settings.NO_EXIT_RUNTIME:
+              logging.warning('you should enable  -s NO_EXIT_RUNTIME=1  so that EVAL_CTORS can work at full efficiency (it gets rid of atexit calls which might disrupt EVAL_CTORS)')
 
       if shared.Settings.ALLOW_MEMORY_GROWTH and shared.Settings.ASM_JS == 1:
         # this is an issue in asm.js, but not wasm
@@ -1935,7 +1941,9 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
         if opt_level >= 2:
           # simplify ifs if it is ok to make the code somewhat unreadable, and unless outlining (simplified ifs
           # with commaified code breaks late aggressive variable elimination)
-          if shared.Settings.SIMPLIFY_IFS and (debug_level == 0 or profiling) and shared.Settings.OUTLINING_LIMIT == 0: JSOptimizer.queue += ['simplifyIfs']
+          # do not do this with binaryen, as commaifying confuses binaryen call type detection (FIXME, in theory, but unimportant)
+          if shared.Settings.SIMPLIFY_IFS and (debug_level == 0 or profiling) and shared.Settings.OUTLINING_LIMIT == 0 and not shared.Settings.BINARYEN:
+            JSOptimizer.queue += ['simplifyIfs']
 
           if shared.Settings.PRECISE_F32: JSOptimizer.queue += ['optimizeFrounds']
 
@@ -1971,7 +1979,7 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           JSOptimizer.flush()
           shared.Building.eliminate_duplicate_funcs(final)
 
-      if shared.Settings.EVAL_CTORS and memory_init_file and debug_level < 4:
+      if shared.Settings.EVAL_CTORS and memory_init_file and debug_level < 4 and not shared.Settings.BINARYEN:
         JSOptimizer.flush()
         shared.Building.eval_ctors(final, memfile)
         if DEBUG: save_intermediate('eval-ctors', 'js')
@@ -2177,7 +2185,10 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
             subprocess.check_call(cmd)
           if import_mem_init:
             # remove and forget about the mem init file in later processing; it does not need to be prefetched in the html, etc.
-            os.unlink(memfile)
+            if DEBUG:
+              safe_move(memfile, os.path.join(emscripten_temp_dir, os.path.basename(memfile)))
+            else:
+              os.unlink(memfile)
             memory_init_file = False
           log_time('asm2wasm')
         if shared.Settings.BINARYEN_PASSES:
@@ -2200,6 +2211,8 @@ There is NO warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR P
           for script in shared.Settings.BINARYEN_SCRIPTS.split(','):
             logging.debug('running binaryen script: ' + script)
             subprocess.check_call([shared.PYTHON, os.path.join(binaryen_scripts, script), final, wasm_text_target], env=script_env)
+        if shared.Settings.EVAL_CTORS:
+          shared.Building.eval_ctors(final, wasm_binary_target, binaryen_bin)
         # after generating the wasm, do some final operations
         if not shared.Settings.WASM_BACKEND:
           if shared.Settings.SIDE_MODULE:
